pyproject.toml
[tool.poetry]
name = "extrator-server"
version = "0.1.0"
description = "Sistema extrator de informações de arquivos pdf."
authors = ["maellson <mqmaellson39@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.11,<4.0"
flask = ">=3.1.0,<4.0.0"
pytesseract = ">=0.3.13,<0.4.0"
pdf2image = ">=1.17.0,<2.0.0"
pymupdf = ">=1.25.5,<2.0.0"
langchain = ">=0.3.24,<0.4.0"
openai = ">=1.76.0,<2.0.0"
faiss-cpu = ">=1.10.0,<2.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
langchain-community = "^0.3.22"
langchain-openai = "^0.3.14"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
